+++ lib/src/gb/cpu.rs
@@ -1,7 +1,7 @@
+use crate::gb::GbModel;
 use crate::gb::bus::CpuBusInterface;
 use crate::gb::ic::ICInterface;
-use crate::gb::GbModel;
-use crate::instructions::{Cond, Instruction, R16Mem, R16Stk, R16, R8};
+use crate::instructions::{Cond, Instruction, R8, R16, R16Mem, R16Stk};
 use crate::utils::bit::{
     add_bytes, add_bytes_carry, add_word_signed_byte, add_words, get_bit, hi, lo,
     rotate_left_get_carry, rotate_left_through_carry, rotate_right_get_carry,
@@ -47,14 +47,14 @@ impl Cpu {
         let flags = if header_checksum == 0x00 {
             Flags {
                 zero: true,
-                subtract: false,
-                half_carry: true,
-                carry: true,
+                ..Default::default()
             }
         } else {
             Flags {
                 zero: true,
-                ..Default::default()
+                subtract: false,
+                half_carry: true,
+                carry: true,
             }
         };